backend/requirements.txt
Django==4.0.7
gunicorn==20.1.0
django-model-utils==4.2.0
djangorestframework==3.13.1
Markdown==3.3.6
django-csp==3.7
psycopg2-binary==2.9.3
coverage==5.5
oauthlib==3.2.0
tblib==1.7.0
email_reply_parser==0.5.12
boto3==1.20.53
django-storages==1.12.3
Pillow==10.2.0
msal==1.17.0
html2text==2020.1.16
PyLTI1p3==2.0.0
